Update patch set 1
Patch Set 1:

(1 comment)

> (5 comments)
 > 
 > Joshua Hesketh wrote:
 > 
 > > In my opinion the buildset resources section of the spec is the
 > most
 > > interesting. The first two sections (Containers That Behave Like
 > a
 > > Machine & Native Container Workflow) are more about doing our
 > > traditional jobs (eg linting) in a container and provide
 > alternative
 > > approaches in implementation.
 > 
 > The second, "Native Container Workflow" is definitely not intended
 > to
 > be like our current jobs.  It's intended to support the workflow of
 > testing an application which runs in containers.
 > 
 > > Using nodepool to hold containers while zuul interacts with them
 > > seems cumbersome and hacky. However, it's perhaps a good solution
 > to
 > > get us to containers without much effort. So if we wanted to
 > reduce
 > > our resource usage by shifting to containers this may be a quick
 > way
 > > of getting there. Otherwise I think we should focus on a more
 > native
 > > container workflow.
 > 
 > This second pargraph makes me think you agree with what I just
 > wrote,
 > but doesn't seem to agree with the first paragraph.  I'm slightly
 > confused.  :)


Heh, it's most likely that I've not understood something.

I think my confusion is that the "Native Container Workflow" talks in terms of building an image and running it. It does not talk about how to interact with the container or what a job might otherwise do. As such I assumed this was more in line of then running tox or similar within the container rather than building an application into a container. In the "BuildSet Resources" section it talks about follow on jobs which I took as building an application into a container as a prerequisite. (This is why I thought the last section was really the interesting part).

 > 
 > > If the job is building the container image and not doing it on a
 > > "full machine from nodepool" then from its current feature set
 > I'm
 > > not sure what purpose Nodepool has? So rather than requiring
 > > nodepool to create an OpenShift project and set up the image
 > > repository etc, perhaps it is worth considering zuul or a new
 > > component to handle that. This will allow us to not arbitrarily
 > > couple nodepool's complexities with container concerns.
 > 
 > I think many of us have seen Nodepool as a "generic resource
 > provider
 > for Zuul" for a while.  There's all kinds of useful things that
 > Zuul
 > could request and Nodepool could provide, including more complex
 > scenarios like "two nodes on a private network".  We've tried to
 > restructure Nodepool so that we can accomodate extensions like this
 > in the future.
 > 

Sure, maybe that's a role filled by nodepool, but I think if we're getting to the point where things are that tightly coupled it might be worth considering making it a part of zuul. This may mean that where it makes sense (specifically simpler/smaller deployments) there is an option to run less serivces. However this is an implementation detail that we can worry about later in my opinion.


 > > Back to buildsets. There should be a job that builds the images.
 > > This is because that's an important part of testing container
 > > workflow. Being able to then run tests against the
 > image/container
 > > is also very useful. Finally, being able to use that image asset
 > in
 > > some way is important. For example, a buildset might: 1) build
 > and
 > > validate the image, 2) run smoke tests, then 3) upload a verified
 > > image to a public repository (ie, more like a release pipeline).
 > 
 > Generally agreed -- but this does touch on an aspect we should
 > probably
 > cover in the spec.  With Zuul, the product of a gate job is still
 > an
 > interim result until it's merged.  So it would not be wise to
 > upload
 > a gate image build to a public repo.  You could do what we do for
 > tarballs, and build the image again in post (trusting that you'll
 > do
 > so with the same commit hashes).  Or you could upload the image to
 > an interim object store, and then promote it to public in a release
 > pipeline.  Either way, publishing a public image extends beyond the
 > gate buildset context.

Unless we can prove ultimate reproducible builds I would say that we need a way for uploading and/or saving what was tested. The gate could propose a Gold Master Candidate and then the post queue could then upload that making it the Gold Master. That said, I feel like this kind of workflow is possible within what this spec produces so long as we have a way of capturing artifacts appropriately. (Then the user could choose where and how it is uploaded). 

 > 
 > > Having the job build the images means that we've lost some of the
 > > scale out gained from slave nodes. In other words, we're making
 > the
 > > zuul-executors do all the work. This might be an argument for
 > doing
 > > the building on a VM from nodepool, but that is also a lot of
 > > overhead for simpler images.
 > 
 > I wouldn't expect the executors to build the image -- if we're not
 > building the image on a vm, then I expect it to be built in a k8s
 > pod, and that shouldn't be running on the executor.


Okay, perhaps this could be clarified a little bit please? Specifically do we have separate pods and nodes (etc) for building vs running? Is it the same supplied pod from nodepool that we build on within a container and then launch back into? (Maybe I'm not understanding k8s/openshift well enough here).

 > 
 > > Additionally we need to consider how we do tests against a
 > > container? Are these simply ran by the job executor again?
 > 
 > I'm starting to think I may have not conveyed something (that's why
 > it's a first draft!).  I expect we can write a job which uses the
 > k8s_raw ansible module to deploy k8s pods based on the images we
 > just built, and then runs tests (however they are implemented)
 > against
 > them.  Those tests could be in their own pod, or they could be,
 > say,
 > a python script running on a VM.  They can be implemented with
 > anything
 > Zuul and Ansible can run.

Right, yes, this is the part that I think I'm stuck on. But maybe that's just me! Thanks for your detailed reply and patience as always :-)

Patch-set: 1
Label: Verified=0