Update patch set 1
Patch Set 1:

Right, it's WIP due to tests which I couldn't get working at the time (attempt not committed).

I had originally had the client do the retries but this got complicated as other threads wait on the job to complete. So the job would get marked as completed, the client would check if successful or if it needed to try again, and then start a new job.

I think that was complex because I hadn't abstracted the job away (it was still a gear.TextJob) so it needed recreating entirely. I should be able to keep this abstraction and move the completion logic back into the client so I'll give that a go.

Patch-set: 1
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0