Reduce test concurrency by 1 cpu
This reduces the default test concurrency by one. We do this by running
`nproc --ignore 1` as the test_run_concurrency command. This Will return
a value >=1 by calculating max(1, nproc - 1).

You can still override the actual concurrency using tox -epy36 --
--concurrency=8 for example.

We do this because the zuul test suite is more of a functional test
suite with running sql and zookeeper databases. We are essentially
setting aside a cpu worth of cpu time for those external services that
the test suite relies on.

Change-Id: Ic515397b2737cf8b9b9ee4dbab0464ac28f6a6c9