Update patch set 13
Patch Set 13: Code-Review+2

(2 comments)

Noted a couple of security-sensitive exceptions inline we may want to have tests to prevent regression. No need to add them to this change in my opinion, could be in a follow-on or something.

Not approving in case Matt still had lingering feedback.

Patch-set: 13
Label: Code-Review=+2