Update patch set 2
Patch Set 2: Code-Review-1

I'd like to have the option to display build times in UTC.  In OpenStack, as a global project, all of our times are in UTC, so when we correlate events across systems, it's better if we don't have to keep mentally converting timezones.

We should definitely store times in the DB in UTC.  I think what you've done there is good.

Since we don't have user preferences for the dashboard, maybe we need a checkbox and cookie to enable the locale conversion?

Or we could consider making this a site-wide setting that the admin sets, and pass the information to the webapp through the info/ endpoint?

Patch-set: 2
Label: Code-Review=-1