Update patch set 1
Patch Set 1:

(1 comment)

[snip]
 > > I think many of us have seen Nodepool as a "generic resource
 > > provider
 > > for Zuul" for a while.  There's all kinds of useful things that
 > > Zuul
 > > could request and Nodepool could provide, including more complex
 > > scenarios like "two nodes on a private network".  We've tried to
 > > restructure Nodepool so that we can accomodate extensions like
 > this
 > > in the future.
 > >
 > 
 > Sure, maybe that's a role filled by nodepool, but I think if we're
 > getting to the point where things are that tightly coupled it might
 > be worth considering making it a part of zuul. This may mean that
 > where it makes sense (specifically simpler/smaller deployments)
 > there is an option to run less serivces. However this is an
 > implementation detail that we can worry about later in my opinion.
 > 

IMO it would be easier and more flexible if Nodepool was part of Zuul already. Having a loose API through Zookeeper may adds complexity to this spec implementation. For example, a k8s driver may want to set node affinity to launch or relaunch job on a node already having the data cached, which sounds like the service would need to somehow be aware of the change it is providing resource too.

Patch-set: 1
Label: Verified=0