WIP: Make config objects freezable
If we want to use configuration objects between layouts, we can't change
them once they are loaded.  To validate that the code doesn't do that,
and prevent regressions, make them immutable after loading is complete.

A variable is added to the base job in the check-vars ansible test to make
sure that we exercise Job.applyVariant in the case where the base job
starts with a variable defined.

Change-Id: I42d01962c4b2dd3cdfce3641ea367b4f4f6892c9