Only emit parent-change-enqueued if needed
When running with GitHub and many app installations emitting the
parent-change-enqueued event is very expensive. In most cases this
event will be used to enqueue downstream changes together with the
parent change into a gate queue. Currently this event is emitted if
any pipeline uses defined trigger regardless if the event will be
consumed or not.

Emitting this event only when there is a real user makes it possible
to use this event in the gate pipeline without overloading github by
depends-on search requests triggered by items enqueued into the check
pipeline.

Change-Id: Iadcfe9b96172dcc3d0affb252a542beed3b90831