Update patch set 3
Patch Set 3: Code-Review+2

There is also datetime.utcfromtimestamp() which might be a little easier to read but meh.

I'm not approving this because it is a change to end user experience. That said I think this is a good improvement and biting this bullet early is probably a good idea.

Patch-set: 3
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2