Remove terminal prompt from commands listed in the docs
The documentation contains commands to install and configure various
tools and services such as zuul, nodepool and gerrit. The commands
are placed in code blocks so it is obvious they are commands to be
executed by the users.

However, the commands contain terminal prompts which make it annyoying
to copy and paste while following the steps. Apart from that, not all
of the commands are preceded with prompts either.

This change removes all the terminal prompts to align the documentation
and make it easier to follow the steps.

Change-Id: I90b571bb4645984f07009b3cb7bd513b9d796d08
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>