Use code-block for admin documentation and remove terminal prompts
The documentation contains commands to install and configure various
tools and services such as zuul, nodepool and gerrit. There are also
yaml snippets as examples.

This change switches to code-blocks for the commands and yaml snippets
to make the documentation consistent. Apart from this, all terminal
prompts are removed for consistency since majority of the documentation
do not use them. This also makes it easier for people who want to follow
steps one by one to do everything manually to get better understanding
regarding how things are put together and copy/paste helps.

Change-Id: I90b571bb4645984f07009b3cb7bd513b9d796d08
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>