Use code-block for admin documentation and terminal prompts
The documentation contains commands to install and configure various
tools and services such as zuul, nodepool, and gerrit. There are also
yaml and ini config file snippets as examples.

This change switches to code-blocks for the commands and snippets to
make the documentation consistent. Apart from this, terminal prompts
are added to comply with OpenStack documentation standards and for
consistency.

Change-Id: I90b571bb4645984f07009b3cb7bd513b9d796d08