Stop deep-copying job variables
This is no longer necessary at parse time.  It's now safe for the
job to maintain a reference to the variable dict within the job
YAML-dict because the Job object itself will be frozen, eliminating
the ability to modify that dict.  When job variants are combined,
rather than modifying any existing variable dictionaries on the Job,
it now creates a new empty dictionary in _deepUpdate, modifies it,
and returns the result.

Change-Id: Ic6f2d4242a98469480e56a266f870dbbd542552b