Add CentOS 7 Nodepool installation instructions
This change adds instructions to install Nodepool on CentOS 7 systems
into Zuul From Scratch document.

Most of the steps are same as Fedora 27 except the package installation
and the use of Nodepool systemd drop-in file. The steps are put under
distro specific headings in order to list the steps more explicitly
and this resulted in duplication. Having separate headings would
be helpful when instructions for Ubuntu 16.04 need to be added so the
duplication might not be a major issue.

Also some of the nodepool configuration steps like creation of the
directories and ssh key generation are moved into nodepool_install.rst
similar to how zuul_install.rst lists some of these steps.

SSH key nodepool_rsa is needed in zuul_install.rst but it gets created later
in nodepool_openstack.rst so moving the key generation to nodepool_install.rst
prevents possible confusion as well.

Depends-On: https://review.openstack.org/#/c/564872/
Change-Id: If073ba5559d3ad055c0c8a1f2f4213ce6ed8af9f