Add CentOS 7 Zuul installation instructions
This change adds instructions to install Zuul on CentOS 7 systems
into Zuul From Scratch document.

Most of the steps are same as Fedora 27 except the package installation
and the use of Zuul systemd drop-in files. The steps are put under
distro specific headings in order to list the steps more explicitly
and this resulted in duplication. Having separate headings would
be helpful when instructions for Ubuntu 16.04 need to be added so the
duplication might not be a major issue.

Depends-On: https://review.openstack.org/#/c/565074/
Change-Id: I830be38cbb52c7b8aa95f1ad080341de40250d17