Add zuul systemd drop-in files for CentOS 7
The existing service files are for Fedora27 and it does not work
on CentOS 7 since pip installs zuul-executor, zuul-scheduler, and
zuul-web in /usr/bin on CentOS 7. This change adds the systemd drop-in
files for this distro so ExecStart and ExecStop commands are overriden
by systemd based on where zuul componenst are installed.

This change is needed in order to have complete installation instructions
for Centos7 in zuul-from-scratch document.

Change-Id: Ib9766ba2a825486f05464be6a85968b5158e2e34