Update patch set 1
Patch Set 1:

> What is the appropriate package under centos 7 then? Doesn't this
 > break some things things?

python34-devel and python36-devel are available on Centos repositories but as you know, python35 is needed. In case of python36-devel, there is no python36-pip. 

That's why the repo ius-release is suggested to be enabled in centos 7 environment prep documentation for python35u-devel python35u, python35u-pip, python35u-devel.

https://review.openstack.org/#/c/564948/4/doc/source/admin/centos7_setup.rst

So this shouldn't break things since it will continue working as intended on Fedora for example.

Patch-set: 1