Update patch set 12
Patch Set 12:

I'm pretty sure we don't need to leave the projects unfrozen, and if so, I'd like to avoid making that change as it seems safer to have them frozen ASAP.  I've pushed up a followup which removes that code to verify.  We may want to either squash it, or merge both changes together.

Patch-set: 12