Update patch set 1
Patch Set 1:

> I think we should add command socket support (like [1]) to zuul-web
 > instead of handling this differently.
 > 
 > [1] https://review.openstack.org/523466

Agree to this but I think it is better to fix the problem as soon as possible since we don't know who might be trying to install zuul and might get confused due to the error message.

The work to add command socket support can be done separately to align it with the rest and the service file can be rolled back to what it is now once the support is there.

Patch-set: 1
Reviewer: Gerrit User 15362 <15362@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0