Fix layout generation for trusted projects
There was a bug introduced by the new handling of non live items in
independent pipeline queues where a trusted project would not find a
proper layout if it depended on another change.

Address this by having the trusted project search backwards through the
entire queue for a valid layout and use the first one found or use the
pipeline layout if not found.

This also adds a test for this case and updates the default
single-tenant config to run tests against common-config, the trusted
project.

Change-Id: I7c0abe2c12423e003a135fa2108a2938d5b0186a