Do not fail so early on missing required-projects
Because of the way the exclude options work in tenant project configs,
we will still parse the job objects, even if we plan to discard them
later. In this case, a missing project is not a big deal. Any jobs that
do require projects that don't exist will fail later down in the process
after they've passed through the exclude filters.

Change-Id: I78b68c5a3c8d403f0837d154a4ac47d11876bae5