Update patch set 1
Patch Set 1:

I will definitely vote for this. I understand you have focused on the admin UI, but the project owner Admin part should be also addressed. i.e. it gives me no benefit if I am not in the tenant admin group. Extension of the project with it's admins is welcome. It's clear, that it will become a mess in the tenant config, might be some different approach?

Patch-set: 1