Update patch set 1
Patch Set 1:

My only concern, that while design addresses lack of possibility to restart job by a project team (without updating branch to re-trigger check/gate) it does this only for limited set of admins. Let's say I am an owner of one project in zuul and need to restart jobs (because of configuration problems or simply lost job) - I can't do this unless I am in this limited admin group (for big installations, like i.e. zuul.openstack it will not be the case, that all project owners are in the admin group). So while you propose a change to give limited admin access to a restricted group it is required to think how/whether it will be possible to extend it in future. What I describe is a very nasty problem, especially if Zuul wants to go as a full standalone CI/CD for use outside of the OpenStack eco.

Patch-set: 1
Reviewer: Gerrit User 27900 <27900@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0