Update patch set 2
Patch Set 2:

So I've posted a follow up patch (mostly as a POC) that implements this as a driver interface[0]. I think that would be a useful improvement but I'm still not sure if it's the right solution.

The challenge is that a tenant may have multiple connections or a connection may be associated with multiple tenants. Really the API should probably be connection specific, ie: /api/{connection}/builds but that would require the user to know about how the databases are deployed.

Somewhat this is a fair requirement though. There may be a particular reason a deployer of zuul has separated out databases. I think for that reason a deployer should perhaps denote which database of results should be presented via the webui (perhaps they didn't want a particular pipeline published etc). We could expose that as an extension of sql_connection_name that is per-tenant.

Alternatively (or perhaps additionally), the driver could expose all connections it knows about (using an interface like in [0]). Basically this solves the problem of multiple connections per tenant as highlighted in the commit message.

What isn't solved in any of these cases is where multiple tenants share a database connection. The results for other tenants will still be exposed unless filtered upon on the 'tenant' column.

[0] https://review.openstack.org/568088

Patch-set: 2
Label: Workflow=0