Ignore extra routes
The way things are setup at the moment, every SQLConnection() object
via getWebHandlers() installs a route for "GET /builds".  When
connections are setup during unit testing for
"resultsdb_<mysql|postresql|mysql_failures|postgresql_failures>" we
end up with 4 attempts to register this.

While this is incorrect, it seems an uncapped aiohttp has become more
strict on this, and raises a RuntimeError now:

 Traceback (most recent call last):
  ..
  File ".../zuul/web/__init__.py", line 401, in run
     app.router.add_route(method, path, handler)
  File ".../.tox/py36/lib/python3.6/site-packages/aiohttp/web_urldispatcher.py", line 857, in add_route
     expect_handler=expect_handler)
  File ".../.tox/py36/lib/python3.6/site-packages/aiohttp/web_urldispatcher.py", line 257, in add_route
     "registered".format(route=route_obj))
  RuntimeError: Added route will never be executed, method GET is already registered

Tracing this check, it seems to be present in our existing versions
[1] (since 1.3.0 even).  So it remains unclear to me what has been
perturbed to trigger this.

This proposes essentially what I belive to be the old behavior of
ignoring the duplicate handlers.  I think the future of this is to
turn into a sensible and helpful error.

[1] https://github.com/aio-libs/aiohttp/commit/4cb301625cb3513a20c0e4c2b8da1b9501101818

Change-Id: Ic90b9014a2e5b5c829d78396ae641b515464440d