Update patch set 2
Patch Set 2:

Yes, I'll push another revision of this patch which implements the tenant filtering.

I think your change is a good one -- I only held back from implementing it that way because it's not clear to be that the builds API (especially by virtue of where it is) should indeed be pluggable.

At any rate, the immediate impetus of this change is to make it so that I can complete the next change -- the cherrypy conversion -- without changing a bunch of code which doesn't quite fit.  Can you base this change on the cherrypy version so we can consider it at the end of the stack?  I'd like to avoid redesigning the driver-webplugin interface in the cherrypy change (which would be necessary if we made this change where it currently is).

Patch-set: 2
Label: Verified=0