github: Do not process comment events for regular issues
Github uses the same webhook event for comments added to both regular
issues and pull requests. Currently, Zuul will attempt to process
non-PR issues which results in a number of API requests as it attempts
to find non-existent pull requests.  This needless traffic can cause
rate-limiting issues in high traffic installations. The webhook body contains
an extra key when comments originate from a PR, which we can use to filter.

Change-Id: I9ab2cc7f37273126cfc669ebf2404215ca6c6f80