Fix new depends-on format matching for prefixed gerrit ui
This patch aims to fix issue where such a depends-on won't match.

Depends-On: https://sftests2.com/r/#/c/1/

Indeed here Gerrit is hosted behind a reverse proxy under /r.
The regexp matching fail. Also the connection source lookup
may fail if Gerrit web interface is not connection.server or
connection.canonical_hostname so this path defines connection.baseurl
as a new lookup condition.

connection.baseurl is then used to build the regexp in order
to set the path prefix of the Gerrit ui.

Change-Id: Ie764bfc814b821636d447f82147a7d76e0511d2f