Update patch set 1
Patch Set 1:

Looking at that I saw that the getRepoPermission method in the FakeGithubConnection shadows the original method in the real GithubConnection. Thus this method is not tested at all. I've a few patches that also move other methods out of the FakeGithubConnection into the FakeGithub and improve testing in that area [1][2]. I might be able to do this for getRepoPermission too within the next two weeks. 

After that we have the ground work to also test that by adding an assert in test_pipeline_require_review_comment_masked.

[1] https://review.openstack.org/#/q/status:open+project:openstack-infra/zuul+branch:master+topic:parent-change-enqueued
[2] https://review.openstack.org/535680

Patch-set: 1