Update patch set 1
Patch Set 1:

Clint, True that we can trigger on incoming recheck/etc but that is for incoming comment events and is handled by a different event-processing code path code path [1]. getPullReviews() is used to collect PullRequest.reviews, which only gets used to determine if a change satisfies the configured review requirements.

Looking closer at zuul v3, it looks like it can now require any type of review, including comments.  I'm not sure of the usefulness of using that, though, since there is no matching of the contents of the comment reviews happening ATM [2]  It looks like there is a TODO around this in the docs [3]  Currently, the only real requirement you can make is that '$someone with $permission left any arbitrary comment on the PR'   
And this patch would break that, so there's also a hole in the GH requirement testing. ;P


[1] https://github.com/openstack-infra/zuul/blob/master/zuul/driver/github/githubconnection.py#L291
[2] https://github.com/openstack-infra/zuul/blob/master/zuul/driver/github/githubmodel.py#L110
[3] https://github.com/openstack-infra/zuul/blob/master/doc/source/admin/drivers/github.rst#requirements-configuration

Patch-set: 1
Reviewer: Gerrit User 1420 <1420@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0