executor: add support for generic build resource
This change adds support for generic build resources to the executor server.
Using the connection_port, kube config file is generated to enable ansible
tasks to use the resource. Context and pod name is also exported in a
zuul_resources variable to enable job to directly use the build resource.

Currently this change supports 'kubectl' resource for containers that behave
like a machine, and 'project'/'namespace' resource for native containers.

Depends-On: https://review.openstack.org/570667
Depends-On: https://review.openstack.org/535557
Change-Id: Icdb9e800177dc770c58f65b02456a6b904be0666