executor: add support for generic build resource
This change adds support for generic build resources to the executor server.
Using the connection_port, access information such as service account and
cluster endpoint are extracted from the node list and injected in a
resources variable file. Currently this change supports 'kubectl' resource
for containers that behave like a machine, and 'project'/'namespace' resource
for the native container workflow.

This change also adds initial support for automatic node's secret setup to
configure the .kube/config file. This can be extended to also setup
ssh keys if/when nodepool will manage ssh access for zuul executor.

Depends-On: https://review.openstack.org/570667
Depends-On: https://review.openstack.org/535557
Change-Id: Icdb9e800177dc770c58f65b02456a6b904be0666