Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I like this.  I think this should be the first of many documents about how we recommend people use zuul.  Because of its current placement, I worry that people may think this is required.  But in fact, it's only even recommended for certain medium-to-large installations (for example, a tenant with a single substantial project may not really need this).

How about we make a new section under the user manual called "Recommended Practices" or "HOWTOs" or something like that, and then put this section under that.  It'll be the only entry now, but I expect we'll add more entries to it soon.

Patch-set: 1
Label: Code-Review=-1