Make files and irrelevant-files matchers exclusive
Files and irrelevant-files will be treated as a pair, so that if
"irrelevant-files" appears, it will erase a previous "files"
attribute.

There are a few situations which can be expressed by using both
matchers.  For example:

    files: tests/*
    irrelevant-files: tests/docs/*

Means run if files in the tests/directory change, unless the only
files which changed are in tests/docs.  This could be expressed
with a single regex on 'files'.  However, when we switch to re2,
that may no longer be the case.

This change sacrifices that expressiveness for the simplicity of
not needing to consider both matchers if they are used, and the
ability to more easily override one with the other (though, a user
could still simply set the unwanted matcher to match nothing or
everything as appropriate).

Change-Id: Ibdae9c3d5ff0ebfc4df843554d548ba984203a4c