Send keepalives on the log callback connection
We set a 5 second timeout on our log callback connection from the
executor to the job node. This has created problems when there are quiet
periods during logging as there may be no data sent on the stream for
more than 5 seconds. Address this by sending keepalives on the socket
every 3 seconds to prevent the socket from being treated as timed out.

Change-Id: I66ab17c55d54a3ae0e3f4572bed72826315b4b22