Fix race in slow start test
There was a race in the slow start test where we could release a build
prior to it entering a waiting state which is a noop. Then the build
enters a waiting state and we never release it.

Fix this by waiting until the build is in a waiting state prior to
releasing the build which ensures we are in a state that release() can
transition out of.

Also add an exception to an earlier check that will catch this case in
the future earlier and with a clearer error message.

These logs illustrate the issue:
http://logs.openstack.org/22/572522/2/check/tox-py36/212f60f/testr_results.html.gz

2018-06-05 22:27:19,113 zuul.test DEBUG Build 2594eadbf07b4052bcd0f41182ce69dd released
2018-06-05 22:27:19,113 zuul.test DEBUG Waiting for <FakeBuild check:test1 1,1> to complete
2018-06-05 22:27:19,119 zuul.test DEBUG Running build 2594eadbf07b4052bcd0f41182ce69dd
2018-06-05 22:27:19,119 zuul.test DEBUG Holding build 2594eadbf07b4052bcd0f41182ce69dd
2018-06-05 22:27:19,119 zuul.test DEBUG Build 2594eadbf07b4052bcd0f41182ce69dd waiting

However, once this race was fixed it became clear there were other races
in checking the executor server's accepting state. I have attempted to
address those by stepping through these state transitions more
explicitly.

* Start by accepting jobs
* test1 forces us to stop accepting new jobs
* On test1 completion increase max jobs to 2 for test2 and test3
* Executor starts accepting jobs again
* Once test2 and test3 are started we stop accepting jobs
* Release test2 and test3 so that they complete
* Check that we have returned to a job accepting state

Change-Id: I4a5858cd1ca37d468f06cb2109bc0e3e9c299f69